fix: IdlImporter: enum value 0xFFFFFFFF (4294967295) now supported

+++ tools/CycloneDDS.Compiler.Common/IdlJson/JsonModels.cs
@@ -169,7 +169,7 @@ namespace CycloneDDS.Compiler.Common.IdlJson
         /// Enum value (for enum members)
         /// </summary>
         [JsonPropertyName("Value")]
-        public int? Value { get; set; }
+        public long? Value { get; set; }
 
         /// <summary>
         /// Member Offset
